services/soar-bridge/requirements.txt
fastapi
uvicorn
requests
pandas
pytz
python-dotenv
pyyaml